+++ src/js/game.jsx
@@ -6,7 +6,6 @@
 
 import React from 'react'
 
-import getGetNextState from './get-get-next-state'
 import replayCodec from './replay'
 import Well from './well.jsx'
 
@@ -21,7 +20,7 @@ class Game extends React.Component {
       bar,
       wellDepth,
       wellWidth,
-      getEnemyAi
+      EnemyAi
     } = this.props
 
     if (rotationSystem.rotations.length < 1) {
@@ -36,12 +35,11 @@ class Game extends React.Component {
       throw Error("Can't have well with width " + String(wellWidth) + ' less than ' + String(minWidth))
     }
 
-    this.getNextState = getGetNextState(rotationSystem, bar, wellDepth, wellWidth)
-    this.enemyAi = getEnemyAi(rotationSystem, this.getNextState, bar, wellDepth, wellWidth)
+    this.enemyAi = EnemyAi(this)
 
     const firstWell = Array(wellDepth).fill(0)
 
-    this.firstState = {
+    this.firstWellState = {
       well: firstWell,
       score: 0,
       piece: rotationSystem.placeNewPiece(wellWidth, this.enemyAi(firstWell))
@@ -66,6 +64,103 @@ class Game extends React.Component {
     this.handleCtrlY = this.handleCtrlY.bind(this)
   }
 
+  /**
+    Input {wellState, piece} and a move, return
+    the new {wellState, piece}.
+  */
+  getNextState (wellState, move) {
+    const {
+      rotationSystem,
+      bar,
+      wellDepth,
+      wellWidth
+    } = this.props
+
+    let nextWell = wellState.well
+    let nextScore = wellState.score
+    let nextPiece = {
+      id: wellState.piece.id,
+      x: wellState.piece.x,
+      y: wellState.piece.y,
+      o: wellState.piece.o
+    }
+
+    // apply transform (very fast now)
+    if (move === 'L') {
+      nextPiece.x--
+    }
+    if (move === 'R') {
+      nextPiece.x++
+    }
+    if (move === 'D') {
+      nextPiece.y++
+    }
+    if (move === 'U') {
+      nextPiece.o = (nextPiece.o + 1) % 4
+    }
+
+    const orientation = rotationSystem.rotations[nextPiece.id][nextPiece.o]
+    const xActual = nextPiece.x + orientation.xMin
+    const yActual = nextPiece.y + orientation.yMin
+
+    if (
+      xActual < 0 || // off left side
+      xActual + orientation.xDim > wellWidth || // off right side
+      yActual < 0 || // off top (??)
+      yActual + orientation.yDim > wellDepth || // off bottom
+      orientation.rows.some((row, y) =>
+        wellState.well[yActual + y] & (row << xActual)
+      ) // obstruction
+    ) {
+      if (move === 'D') {
+        // Lock piece
+        nextWell = wellState.well.slice()
+
+        const orientation = rotationSystem.rotations[wellState.piece.id][wellState.piece.o]
+
+        // this is the top left point in the bounding box of this orientation of this piece
+        const xActual = wellState.piece.x + orientation.xMin
+        const yActual = wellState.piece.y + orientation.yMin
+
+        // row by row bitwise line alteration
+        for (let row = 0; row < orientation.yDim; row++) {
+          // can't negative bit-shift, but alas X can be negative
+          nextWell[yActual + row] |= (orientation.rows[row] << xActual)
+        }
+
+        // check for complete lines now
+        // NOTE: completed lines don't count if you've lost
+        for (let row = 0; row < orientation.yDim; row++) {
+          if (
+            yActual >= bar &&
+            nextWell[yActual + row] === (1 << wellWidth) - 1
+          ) {
+            // move all lines above this point down
+            for (let k = yActual + row; k > 1; k--) {
+              nextWell[k] = nextWell[k - 1]
+            }
+
+            // insert a new blank line at the top
+            // though of course the top line will always be blank anyway
+            nextWell[0] = 0
+
+            nextScore++
+          }
+        }
+        nextPiece = null
+      } else {
+        // No move
+        nextPiece = wellState.piece
+      }
+    }
+
+    return {
+      well: nextWell,
+      score: nextScore,
+      piece: nextPiece
+    }
+  }
+
   handleClickStart () {
     const {
       replayTimeoutId
@@ -81,7 +176,7 @@ class Game extends React.Component {
     this.setState({
       mode: 'PLAYING',
       wellStateId: 0,
-      wellStates: [this.firstState],
+      wellStates: [this.firstWellState],
       replay: [],
       replayTimeoutId: undefined
     })
@@ -115,7 +210,7 @@ class Game extends React.Component {
     this.setState({
       mode: 'REPLAYING',
       wellStateId: wellStateId,
-      wellStates: [this.firstState],
+      wellStates: [this.firstWellState],
       replay: replay,
       replayTimeoutId: nextReplayTimeoutId
     })
