--- src/js/moves.js
@@ -1,3 +1,0 @@
-'use strict'
-
-export default ['L', 'R', 'D', 'U']
